+++ changelog.md
@@ -107,3 +107,7 @@
 
 #1.0.24
 * Replaced carets (^) in package with tildes (~)
+
+#1.0.25
+* Improved step options "targetElement" to allow string selector (lazy loaded elements)
+* Added step option for "customPopperContent", to pass an elementRef